Updated code standards. Refactored classes.
