+++ XnaDarts/XnaDarts.sln
@@ -1,13 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+VisualStudioVersion = 12.0.40629.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XnaDartsTests", "XnaDartsTests\XnaDartsTests.csproj", "{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XnaDarts", "XnaDarts\XnaDarts\XnaDarts.csproj", "{8030CF46-66EA-426D-A303-55E3A574953E}"
 EndProject
-Project("{96E2B04D-8817-42C6-938A-82C39BA4D311}") = "XnaDartsContent", "XnaDarts\XnaDartsContent\XnaDartsContent.contentproj", "{403C79B6-8EB8-49FA-A143-F7795391A7CB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XnaDartsContent", "XnaDarts\XnaDartsContent\XnaDartsContent.contentproj", "{403C79B6-8EB8-49FA-A143-F7795391A7CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -18,12 +18,14 @@ Global
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}.Debug|Any CPU.Build.0 = Debug|x86
 		{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}.Debug|x86.ActiveCfg = Debug|x86
 		{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}.Debug|x86.Build.0 = Debug|x86
 		{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}.Release|Any CPU.ActiveCfg = Release|x86
 		{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}.Release|x86.ActiveCfg = Release|x86
 		{1234F2B9-4D3F-4686-9E33-5AB147BFEAF6}.Release|x86.Build.0 = Release|x86
 		{8030CF46-66EA-426D-A303-55E3A574953E}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{8030CF46-66EA-426D-A303-55E3A574953E}.Debug|Any CPU.Build.0 = Debug|x86
 		{8030CF46-66EA-426D-A303-55E3A574953E}.Debug|x86.ActiveCfg = Debug|x86
 		{8030CF46-66EA-426D-A303-55E3A574953E}.Debug|x86.Build.0 = Debug|x86
 		{8030CF46-66EA-426D-A303-55E3A574953E}.Release|Any CPU.ActiveCfg = Release|x86
