+++ XnaDarts/XnaDarts/XnaDarts/Gameplay/Modes/Record.cs
@@ -1,0 +1,22 @@
+using System;
+
+namespace XnaDarts.Gameplay.Modes
+{
+    [Serializable]
+    public struct Record
+    {
+        public DateTime Date;
+        public int Score;
+
+        public Record(int score, DateTime date)
+        {
+            Score = score;
+            Date = date;
+        }
+
+        public override string ToString()
+        {
+            return Score + "          " + Date;
+        }
+    }
+}
